Use dataplane flow plugin protobuf definition directly
Instead of assuming the plugin provides language bindings for it's
protobug message use the definition directly.
This frees the plugin to not have to provide (and thus support)
specific bindings. For example there are at least 3 perl protobuf
bindings (none of which are "offical"), and the plugin's not know
which a clients might want to use. The same applies to
Go/Python, (although at for Go there are "official" protobuf bindings)
to which the client is likely to switch when it's ported to VCI.

Fixes: VRVDR-50591